Move survey responses into separate collection
