+++ src/main/java/network/venox/cobalt/MongoProvider.java
@@ -24,7 +24,7 @@ public class MongoProvider {
         final String url = library.fileSettings.file.yaml.node("mongo").getString();
         if (url == null) throw new IllegalArgumentException("MongoDB URL not found in config!");
         database = new SingleMongo(url).database.loadMagicCollections(MapGenerator.HASH_MAP.mapOf(
-                List.of("auto_slowmodes", "auto_threads", "corners", "corner_creators", "users", "limited_messages", "locks", "react_channels", "servers", "sticky_messages", "surveys", "thread_messages"),
-                List.of(AutoSlowmode.class, AutoThread.class, Corner.class, CornerCreator.class, CoUser.class, LimitedMessages.class, Lock.class, ReactChannel.class, Server.class, StickyMessage.class, Survey.class, ThreadMessage.class)));
+                List.of("auto_slowmodes", "auto_threads", "corners", "corner_creators", "users", "limited_messages", "locks", "react_channels", "servers", "sticky_messages", "surveys", "survey_responses", "thread_messages"),
+                List.of(AutoSlowmode.class, AutoThread.class, Corner.class, CornerCreator.class, CoUser.class, LimitedMessages.class, Lock.class, ReactChannel.class, Server.class, StickyMessage.class, Survey.class, SurveyResponse.class, ThreadMessage.class)));
     }
 }
